requirements.txt
numpy>=1.24.0
open3d>=0.17.0
scikit-learn>=0.24.0
torch>=2.0.0
torchvision>=0.10.0
Pillow>=8.3.1
opencv-python>=4.8.0
pycocotools>=2.0.0
tqdm>=4.65.0
ultralytics>=8.0.0
PyYAML>=6.0
matplotlib>=3.7.0
scipy>=1.10.0 